First version of Viability chart
Added the first version of viability chart.
Y- axis is not currently based on anything ( economics not yet implemented). Also the viability chart currently shows generation based on a day based on the month that was selected in the Grid-size step. Will be changed to yearly generation in the near future.